Adds more tests
But mainly tries to address #22 pinpointing the origin of the problem, which is that @globbers value is nil. This is why I have added some tests, and in order to do that, exported simplify, which helps in that.

I think that the problem is somewhere along the lines of creatting globbers, but without extensive testing of the grammars and base classes, it's difficult to dind it.